+++ HoneyDrunk.Notify/HoneyDrunk.Notify.Abstractions/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Aligned package version to `0.3.0` and Kernel abstractions dependency to `0.7.0`.
+
 ## [0.2.0] - 2026-05-05
 
 ### Changed